refactor: 将 Adder 移出无用目录
